+++ NEWS.md
@@ -8,6 +8,8 @@ has been reorganized and has a lot of useful new content thanks to Martin Modrá
 * The [LOO predictive checks](http://mc-stan.org/bayesplot/reference/PPC-loo.html) 
 now require **loo** version `>= 2.0.0`. (#139)
 
+* Histogram plots gain a `breaks` argument that can be used as an alternative to `binwidth`. (#148)
+
 * [`mcmc_pairs()`](http://mc-stan.org/bayesplot/reference/MCMC-scatterplots.html) 
 now has an argument `grid_args` to provide a way of passing optional arguments to 
 `gridExtra::arrangeGrob()`. This can be used to add a title to the plot, for example. (#143)
@@ -16,8 +18,6 @@ now has an argument `grid_args` to provide a way of passing optional arguments t
 gains an argument `discrete`, which is `FALSE` by default, but can be used to make the 
 Geom more appropriate for discrete data. (#145)
 
-* [`mcmc_hist()`](http://mc-stan.org/bayesplot/reference/MCMC-distributions.html) 
-gains a `breaks` argument that can be used as an alternative to `binwidth`.
 
 # bayesplot 1.5.0